fix: Alterando tipo de dado da varíavel valor de Double para BigDecimal.

+++ src/main/java/com/example/carteiradigital/dto/PixTransferDTO.java
@@ -3,10 +3,12 @@ package com.example.carteiradigital.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class PixTransferDTO {
     private Long origemId;
     private Long destinoId;
-    private double valor;
+    private BigDecimal valor;
 }
